+++ crates/phonon-kira/src/lib.rs
@@ -1,3 +1,3 @@
-pub mod direct_effect;
+pub mod spatializer_effect;
 pub mod eq_effect;
 pub mod reverb_effect;
